- Refactored the decoder to improve code readability - Renamed the decoder to avoid conflicts - Added basic CLI interface
